+++ src/ui/operation-modes/modes/fxcol-editor/Context.cpp
@@ -1,5 +1,6 @@
 #include "Context.h"
 #include <resources/managed-memory/ManagedCArray.h>
+#include <utilities/BoundingBoxes.h>
 #include <algorithm>
 
 namespace ui::operation_modes::modes::fxcol_editor {
@@ -16,6 +17,8 @@ namespace ui::operation_modes::modes::fxcol_editor {
 		for (auto& volume : boundingVolumes)
 			if (shapeIdx >= volume.shapeStartIdx && shapeIdx < volume.shapeStartIdx + static_cast<int>(volume.shapeCount))
 				return &volume;
+
+		return nullptr;
 	}
 
 	void Context::AddBoundingVolume() {
@@ -119,7 +122,77 @@ namespace ui::operation_modes::modes::fxcol_editor {
 		auto* fxColManager = hh::game::GameManager::GetInstance()->GetService<FxColManager>();
 		auto* fxColData = fxColManager->resource->fxColData;
 
-		boundingVolume->aabbMin = { -INFINITY, -INFINITY, -INFINITY };
-		boundingVolume->aabbMax = { INFINITY, INFINITY, INFINITY };
+		if (boundingVolume->shapeCount == 0) {
+			// I prefer to set it as inverse infinities but this is what the game initializes an empty aabb to.
+			boundingVolume->aabbMin = { 0.0f, 0.0f, 0.0f };
+			boundingVolume->aabbMax = { 0.0f, 0.0f, 0.0f };
+		}
+		else {
+			csl::geom::Aabb aabb{ { INFINITY, INFINITY, INFINITY }, { -INFINITY, -INFINITY, -INFINITY } };
+
+			for (int i = boundingVolume->shapeStartIdx; i < boundingVolume->shapeStartIdx + boundingVolume->shapeCount; i++)
+				AddAabb(aabb, CalculateAabb(&fxColData->collisionShapes[i]));
+
+			boundingVolume->aabbMin = aabb.min;
+			boundingVolume->aabbMax = aabb.max;
+		}
+
+		RecalculateKdTree();
+	}
+
+	void Context::RecalculateKdTree() {
+		auto* fxColManager = hh::game::GameManager::GetInstance()->GetService<FxColManager>();
+
+		auto* oldResource = fxColManager->collisionTreeResource;
+
+		auto* newTree = fxColManager->CreateKdTree(fxColManager->resource);
+		auto* newResource = static_cast<hh::ut::ResKdTreeData*>(fxColManager->GetAllocator()->Alloc(hh::ut::KdTreeUtil::CalcMemorySize(*newTree), 16));
+		hh::ut::KdTreeUtil::BuildResource(*newTree, newResource);
+
+		fxColManager->collisionTree = newTree;
+		fxColManager->collisionTreeResource = newResource;
+
+		fxColManager->GetAllocator()->Free(oldResource);
+	}
+
+	csl::geom::Aabb Context::CalculateAabbForObb(const csl::math::Position& position, const csl::math::Rotation& rotation, const csl::math::Vector3& halfExtents) {
+		csl::geom::Aabb aabb{ { INFINITY, INFINITY, INFINITY }, { -INFINITY, -INFINITY, -INFINITY } };
+		Eigen::Affine3f transform{ Eigen::Translation3f{ position } *rotation };
+
+		aabb.AddPoint(transform * -halfExtents);
+		aabb.AddPoint(transform * halfExtents);
+		aabb.AddPoint(transform * -csl::math::Vector3{ halfExtents.x(), halfExtents.y(), -halfExtents.z() });
+		aabb.AddPoint(transform * csl::math::Vector3{ halfExtents.x(), halfExtents.y(), -halfExtents.z() });
+		aabb.AddPoint(transform * -csl::math::Vector3{ halfExtents.x(), -halfExtents.y(), halfExtents.z() });
+		aabb.AddPoint(transform * csl::math::Vector3{ halfExtents.x(), -halfExtents.y(), halfExtents.z() });
+		aabb.AddPoint(transform * -csl::math::Vector3{ halfExtents.x(), -halfExtents.y(), -halfExtents.z() });
+		aabb.AddPoint(transform * csl::math::Vector3{ halfExtents.x(), -halfExtents.y(), -halfExtents.z() });
+
+		return aabb;
+	}
+
+	csl::geom::Aabb Context::CalculateAabb(const app::gfx::FxColCollisionShape* shape) {
+		switch (shape->shape) {
+		case FxColCollisionShape::Shape::SPHERE: {
+			csl::geom::Aabb aabb{ { INFINITY, INFINITY, INFINITY }, { -INFINITY, -INFINITY, -INFINITY } };
+			auto& extents = shape->extents.sphere;
+			aabb.AddPoint(shape->position + csl::math::Vector3{ extents.radius, extents.radius, extents.radius });
+			aabb.AddPoint(shape->position - csl::math::Vector3{ extents.radius, extents.radius, extents.radius });
+			return aabb;
+		}
+		case FxColCollisionShape::Shape::CYLINDER: {
+			auto& extents = shape->extents.cylinder;
+			return CalculateAabbForObb(shape->position, shape->rotation, { extents.radius, extents.halfHeight, extents.radius });
+		}
+		case FxColCollisionShape::Shape::ANISOTROPIC_OBB: {
+			auto& extents = shape->extents.anisotropicObb;
+			return CalculateAabbForObb(shape->position, shape->rotation, { extents.width / 2, extents.height / 2, extents.depth / 2 });
+			break;
+		}
+		case FxColCollisionShape::Shape::ISOTROPIC_OBB: {
+			auto& extents = shape->extents.isotropicObb;
+			return CalculateAabbForObb(shape->position, shape->rotation, { extents.width / 2, extents.height / 2, extents.depth / 2 });
+		}
+		}
 	}
 }