+++ src/ui/operation-modes/modes/level-editor/LevelEditor.cpp
@@ -88,9 +88,6 @@ namespace ui::operation_modes::modes::level_editor {
 		case StopChangingParametersAction::id:
 			GetContext().RespawnActiveObject(GetBehavior<SelectionBehavior<Context>>()->GetSelection()[0]);
 			break;
-		case PlacementBehavior<Context>::ObjectPlacedAction::id:
-			GetBehavior<SelectionBehavior<Context>>()->Select(static_cast<const PlacementBehavior<Context>::ObjectPlacedAction&>(action).payload);
-			break;
 		case FocusObjectDataAction::id:
 			GetBehavior<SelectionBehavior<Context>>()->Select(static_cast<const FocusObjectDataAction&>(action).payload);
 			break;